"Pixels per meter" using float update
Added build script
Fixed unregistering class causing error on build
Updated pixels per meter from int to float
Updated README to reflect changes

+++ modules/sprite_sheet_maker_utils.py
@@ -52,7 +52,7 @@ class AutoCaptureParam:
         self.objects:set = set({})
         self.consider_armature_bones:bool = False
         self.camera_direction:CameraDirection = CameraDirection.NEG_X
-        self.pixels_per_meter:int = 500
+        self.pixels_per_meter:float = 500.0
         self.camera_padding:float = 0.05
 
 class PixelateParam:
